+++ packages/config/src/index.ts
@@ -1,3 +1,7 @@
 export { loadNocturneConfig, NocturneConfig } from "./config";
-export { NocturneContractDeployment } from "./deployment";
+export {
+  NocturneContractDeployment,
+  ProtocolAllowlist,
+  ProtocolWhitelistEntry,
+} from "./deployment";
 export * from "./proxy";