+++ packages/deploy/index.ts
@@ -1,8 +1,4 @@
-export {
-  NocturneDeployConfig,
-  NocturneDeployOpts,
-  ProtocolWhitelistEntry,
-} from "./src/config";
+export { NocturneDeployConfig, NocturneDeployOpts } from "./src/config";
 export { deployNocturne } from "./src/deploy";
 export { whitelistProtocols } from "./src/whitelist";
 export { checkNocturneContractDeployment } from "./src/checks";