+++ packages/deploy/scripts/deploy.ts
@@ -5,6 +5,8 @@ import * as dotenv from "dotenv";
 import * as fs from "fs";
 import { checkNocturneContractDeployment } from "../src/checks";
 import * as JSON from "bigint-json-serialization";
+import { Handler__factory } from "@nocturne-xyz/contracts";
+import { whitelistProtocols } from "../src/whitelist";
 
 const CONFIGS_DIR = `${__dirname}/../configs/`;
 const DEPLOYS_DIR = `${__dirname}/../deploys/`;
@@ -30,11 +32,22 @@ dotenv.config();
   );
   const config: NocturneDeployConfig = JSON.parse(configString);
 
+  // Deploy contracts
   const deployment = await deployNocturne(deployer, config);
-
-  console.log(deployment);
-
   await checkNocturneContractDeployment(deployment, provider);
+
+  // Whitelist protocols
+  const handler = Handler__factory.connect(
+    deployment.handlerProxy.proxy,
+    deployer
+  );
+  await whitelistProtocols(deployer, config.protocolAllowlist, handler);
+
+  const deploymentAndAllowlist = {
+    deployment,
+    protocolAllowlist: config.protocolAllowlist,
+  };
+  console.log(deploymentAndAllowlist);
 
   if (!fs.existsSync(DEPLOYS_DIR)) {
     fs.mkdirSync(DEPLOYS_DIR);
@@ -42,7 +55,7 @@ dotenv.config();
 
   fs.writeFileSync(
     `${DEPLOYS_DIR}/${deployment.network.name}-${Date.now().toString()}.json`,
-    JSON.stringify(deployment),
+    JSON.stringify(deploymentAndAllowlist),
     {
       encoding: "utf8",
       flag: "w",