+++ packages/deploy/src/config.ts
@@ -1,11 +1,13 @@
-import { ProxyAdmin } from "@nocturne-xyz/contracts";
+import { ProtocolAllowlist } from "@nocturne-xyz/config";
 import { Address } from "./utils";
+import { ProxyAdmin } from "@nocturne-xyz/contracts";
 
 export interface NocturneDeployConfig {
   proxyAdminOwner: Address;
   screeners: Address[];
   subtreeBatchFillers: Address[];
   wethAddress: Address;
+  protocolAllowlist: ProtocolAllowlist;
   opts?: NocturneDeployOpts;
 }
 
@@ -13,11 +15,4 @@ export interface NocturneDeployOpts {
   proxyAdmin?: ProxyAdmin;
   useMockSubtreeUpdateVerifier?: boolean;
   confirmations?: number;
-}
-
-export type ProtocolWhitelistConfig = Map<string, ProtocolWhitelistEntry>;
-
-export interface ProtocolWhitelistEntry {
-  contractAddress: Address;
-  functionSignatures: string[];
 }