+++ packages/e2e-tests/src/deploy.ts
@@ -307,6 +307,7 @@ export async function deployContractsWithDummyAdmins(
     screeners: args.screeners,
     subtreeBatchFillers: args.subtreeBatchFillers,
     wethAddress: weth.address,
+    protocolAllowlist: new Map(),
     opts: {
       useMockSubtreeUpdateVerifier:
         process.env.ACTUALLY_PROVE_SUBTREE_UPDATE == undefined,
@@ -388,6 +389,7 @@ export async function setupTestClient(
 ): Promise<ClientSetup> {
   const config = new NocturneConfig(
     contractDeployment,
+    new Map(), // dummy value, don't need whitelist here
     // TODO: fill with real assets and rate limits in SDK gas asset and deposit
     // screener PRs
     opts?.gasAssets ?? new Map(Object.entries({})),