+++ packages/config/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### Unreleased
 
+- add `protocolAllowlist` to NocturneConfig
 - add `toString`, `fromString` methods to avoid direct map serialization
-- add `fromObject` method 
+- add `fromObject` method
 - Add `handler` (remove vault) and add `handlerOwner`
 - Add checks in single unit test to ensure fields are defined
 - Add `depositManager`, `depositSources`, `screeners`, and `owners` (added `walletOwner` and `depositManagerOwner` fields)